Move the passes into a namespace, so that the parsing namespace has just the core API.
