+++ src/Core/Gibberish.Tests/RecognizeBlockSyntax/InterpretWholeFile.cs
@@ -1,5 +1,6 @@
 ﻿using Gibberish.AST._1_Bare;
 using Gibberish.Parsing;
+using Gibberish.Parsing.Passes;
 using Gibberish.Tests.ZzTestHelpers;
 using NUnit.Framework;
