+++ src/Core/Gibberish/Gibberish.csproj
@@ -97,7 +97,7 @@
     <Compile Include="AST\Parse.cs" />
     <Compile Include="AST\PassStatement.cs" />
     <Compile Include="AST\_1_Bare\UnknownStatement.cs" />
-    <Compile Include="Parsing\AssembleBlocks.cs" />
+    <Compile Include="Parsing\Passes\AssembleBlocks.cs" />
     <Compile Include="AST\_1_Bare\LanguageConstructVisitor.cs" />
     <Compile Include="AST\ParseError.cs" />
     <Compile Include="AST\_1_Bare\CommentDefinitionBlock.cs" />
@@ -110,8 +110,8 @@
     <Compile Include="Parsing\LineParsers\ParseStatements.cs" />
     <Compile Include="Parsing\ParsingExtensions.cs" />
     <Compile Include="AST\_1_Bare\UnknownPrelude.cs" />
-    <Compile Include="Parsing\RecognizeLines.cs" />
-    <Compile Include="Parsing\Transform.cs" />
+    <Compile Include="Parsing\Passes\RecognizeLines.cs" />
+    <Compile Include="Parsing\Passes\Transform.cs" />
     <Compile Include="Parsing\UnfixableError.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="AST\Statement.cs" />
